Changed TokenBase::compile from (Program &, x86::Gp *) to (Program &, regdefp_t) where:  typedef std::pair<asmjit::x86::Gp *, DataDef **> regdefp_t
All tokens contain an x86::Gp and a DataDef *, so they have access to their own
"general purpose" register and data type. When their compile() method is called
they can choose to use the register provided to store data in versus their own,
and also can use the provided data type for typecasting purposes.

This should allow for the TokenDiv in the following line to do an IDIV
    int x = 10 / 2;
and for the TokenDiv in the following line to do an FDIV
    double x = 10 / 3;

because the DataDef pointer from the variable "x" would be given to the
TokenDiv.compile by TokenAssign.compile